chore(version): cosmeticly reorganize the version crate
